L2_order_book_project/requirements_L2_order_book_project.txt
pandas
websocket-client
pyarrow
pyyaml
numpy
scikit-learn
xgboost
